refactor: 장바구니관련 통신 라이브러리 Retrofit으로 변경

+++ app/src/main/java/woowacourse/shopping/feature/cart/CartContract.kt
@@ -24,7 +24,7 @@ interface CartContract {
         fun minusPageNumber()
         fun plusQuantity(cartProductState: CartProductState)
         fun minusQuantity(cartProductState: CartProductState)
-        fun updateChecked(productId: Int, checked: Boolean)
+        fun updateChecked(productId: Long, checked: Boolean)
         fun deleteCartProduct(cartProductState: CartProductState)
         fun checkAll()
     }
